# improve performance, alternative approach (+149% throughput and more)