@sindresorhus 2023-03-24
Thanks :)