@sindresorhus 2023-03-24
https://github.com/chalk/slice-ansi/releases/tag/v6.0.0